refactor(core): slim main to its state, and document the performance contract
main.rs drops from 1,443 lines to ~500: the boot sequence (panic hook
through terminal init) moves to coordinator/boot.rs, the terminal/error-
panel/perf-HUD helpers to coordinator/chrome.rs, both within ADR-22's
grouping — App, its field docs and the loop constants stay put and main()
is one delegation. Terminals::new stops hand-building the backend
registry across five agent crossings: BackendRegistry::
from_configured_hosts owns how backends are registered, lazily as
before. SpawnRequest derives Default, so its seven call sites state only
what they mean. The theme-background pass walks the buffer's contiguous
storage like the width normalisation already does.

Docs catch up with the performance work as one contract: ADR-P18 in
docs/PERFORMANCE.md records the closure of ADR-P16's volatile-group
list, the cache-age rule going uniform, and what got cheaper to ask of
the world; CLAUDE.md's render-loop and remote sections say the same
where they said the old state. And the guidance for writing FAST custom
panes now ships with the interface: the levers (pure, identity-keyed
memoization, window-first, hoisting, the shared spinner clock, the perf
HUD as the check) are in docs/PLUGINS.md, ui/AGENTS.md — the file a
coding CLI loads when editing the interface — and the thurbox-ui skill.
